chore: extact useSidePanel hook [YTFRONT-4463]
